+++ django_vite/__init__.py
@@ -1,4 +1,4 @@
-from .loader import DjangoViteConfig
+from .core.asset_loader import DjangoViteConfig
 
 
 __all__ = ["DjangoViteConfig"]
