+++ django_vite/apps.py
@@ -3,7 +3,7 @@ from contextlib import suppress
 from django.apps import AppConfig
 from django.core.checks import Warning, register
 
-from .exceptions import DjangoViteManifestError
+from .core.exceptions import DjangoViteManifestError
 from .templatetags.django_vite import DjangoViteAssetLoader
