+++ django_vite/core/exceptions.py
